:muscle: modify the adding polls functionality
achieve adding of poll choices in main polls table. this is by modifying the code in the pollsController to change data received from the json as an array into a CSV before sending it to the database. this worked well with the only drawback being that you cannot pass choices with commas within as the code would break.