:ferris_wheel: modify the change an forgot password logic.
if a user forgets their password, they input their email after which  a validation link will be sent to it and upon clicking redirected to a page to reset it. Same as on in which a user already logged in can access. on it they will input the new password and confirm and having done so will have updated their password.